Server/requirements.txt
Flask
pandas
scikit-learn
Flask-Cors
python-dotenv